# swagger/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Categorie",
      "description": "Operazioni sulle categorie"
    },
    {
      "name": "Configurazioni",
      "description": "Operazioni sulle configurazioni"
    }
  ],
  "paths": {
    "/configurations": {
      "post": {
        "tags": [
          "Configurazioni"
        ],
        "summary": "Crea o aggiorna una configurazione",
        "description": "Crea o aggiorna la configurazione per uno store e una source specifici.",
        "operationId": "createConfiguration",
        "parameters": [
          {
            "name": "X-STORE",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-SOURCE",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configurazione creata/aggiornata con successo",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoString"
                }
              }
            }
          },
          "400": {
            "description": "Richiesta non valida",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoString"
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categorie"
        ],
        "summary": "Recupera le categorie",
        "description": "Restituisce la lista delle categorie disponibili per uno store e una source specifici.",
        "operationId": "getCategories",
        "parameters": [
          {
            "name": "X-STORE",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-SOURCE",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successo",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListCategoryDto"
                }
              }
            }
          },
          "401": {
            "description": "Non autorizzato",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListCategoryDto"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListCategoryDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConfigurationRequest": {
        "type": "object",
        "properties": {
          "apiToken": {
            "type": "string"
          },
          "shopId": {
            "type": "string"
          }
        },
        "required": [
          "apiToken",
          "shopId"
        ]
      },
      "ResponseDtoString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          }
        }
      },
      "ResponseDtoListCategoryDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            }
          }
        }
      }
    }
  }
}